+++ src/service/payment/chevre.ts
@@ -47,9 +47,16 @@ export function authorize(params: {
         const transaction = await repos.transaction.findInProgressById({ typeOf: params.purpose.typeOf, id: params.purpose.id });
 
         const paymentServiceType = params.paymentServiceType;
-        // プロジェクトの対応決済サービスを確認
-        // const paymentMethodType = params.object.paymentMethod;
-        // const paymentServiceType = await getPaymentServiceType({ project: { id: params.project.id }, paymentMethodType });
+
+        // if (paymentServiceType === chevre.factory.service.paymentService.PaymentServiceType.Account) {
+        //     // プロジェクトの対応決済サービスを確認(Account->PaymentCard移行のため)
+        //     const paymentMethodType = params.object.paymentMethod;
+        //     const paymentServiceTypeByPaymentMethodType
+        //         = await getPaymentServiceType({ project: { id: params.project.id }, paymentMethodType });
+        //     if (typeof paymentServiceTypeByPaymentMethodType === 'string') {
+        //         paymentServiceType = paymentServiceTypeByPaymentMethodType;
+        //     }
+        // }
 
         // 取引番号生成
         const transactionNumberService = new chevre.service.TransactionNumber({
@@ -332,7 +339,8 @@ async function getPaymentServiceType(params: {
         typeOf: {
             $in: [
                 chevre.factory.service.paymentService.PaymentServiceType.CreditCard,
-                chevre.factory.service.paymentService.PaymentServiceType.MovieTicket
+                chevre.factory.service.paymentService.PaymentServiceType.MovieTicket,
+                chevre.factory.service.paymentService.PaymentServiceType.PaymentCard
             ]
         },
         serviceOutput: { typeOf: { $eq: params.paymentMethodType } }