+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - AccountプロダクトとPaymentCardプロダクトの挙動を統一
+- 口座注文時に、口座にアクセスコードを設定
 
 ### Deprecated