+++ CHANGELOG.md
@@ -18,6 +18,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v10.9.0 - 2020-11-17
+
+### Changed
+
+- AccountプロダクトタイプをPaymentCardに統合
+- 口座注文時に、口座にアクセスコードを設定
+
 ## v10.8.0 - 2020-11-13
 
 ### Changed